fixing cpu-only mode
