+++ dali/operators/util/get_property.cc
@@ -39,118 +39,6 @@ The type of the output will depend on the ``key`` of the requested property.)cod
 )code",
             DALI_STRING);
 
-
 DALI_REGISTER_OPERATOR(GetProperty, GetProperty<CPUBackend>, CPU)
-DALI_REGISTER_OPERATOR(GetProperty, GetProperty<GPUBackend>, GPU)
-
-namespace detail {
-namespace {
-
-const DALIMeta& GetMeta(const TensorVector<CPUBackend>& batch, int tensor_idx) {
-  return batch[tensor_idx].GetMeta();
-}
-
-const DALIMeta& GetMeta(const TensorList<GPUBackend>& batch, int tensor_idx) {
-  return batch.GetMeta(static_cast<int>(tensor_idx));
-}
-
-template <typename Backend, typename BatchContainer = batch_container_t<Backend>>
-struct SourceInfo : public Property<Backend> {
-  TensorListShape<> GetShape(const BatchContainer& input) override {
-    TensorListShape<> ret{static_cast<int>(input.num_samples()), 1};
-    for (int i = 0; i < ret.size(); i++) {
-      ret.set_tensor_shape(i, {static_cast<int64_t>(GetSourceInfo(input, i).length())});
-    }
-    return ret;
-  }
-
-  DALIDataType GetType(const BatchContainer&) override {
-    return DALI_UINT8;
-  }
-
-  void FillOutput(workspace_t<Backend>& ws) override {
-    const auto& input = ws.template Input<Backend>(0);
-    auto& output = ws.template Output<Backend>(0);
-    for (size_t sample_id = 0; sample_id < input.num_samples(); sample_id++) {
-      auto si = GetSourceInfo(input, sample_id);
-      output[sample_id].Copy(make_cspan((const uint8_t*)si.c_str(), si.length()), nullptr);
-    }
-  }
-
- private:
-  const std::string& GetSourceInfo(const BatchContainer& input, size_t idx) {
-    return GetMeta(input, idx).GetSourceInfo();
-  }
-};
-
-template <>
-void SourceInfo<GPUBackend>::FillOutput(workspace_t<GPUBackend>& ws) {
-  const auto& input = ws.template Input<GPUBackend>(0);
-  auto& output = ws.template Output<GPUBackend>(0);
-  for (size_t sample_id = 0; sample_id < input.num_samples(); sample_id++) {
-    auto si = GetSourceInfo(input, sample_id);
-    auto output_ptr = output.raw_mutable_tensor(static_cast<int>(sample_id));
-    cudaMemcpyAsync(output_ptr, si.c_str(), si.length(), cudaMemcpyDefault, ws.stream());
-  }
-}
-
-
-template <typename Backend, typename BatchContainer = batch_container_t<Backend>>
-struct Layout : public Property<Backend> {
-  TensorListShape<> GetShape(const BatchContainer& input) override {
-    // Every tensor in the output has the same number of dimensions
-    return uniform_list_shape(input.num_samples(), {GetLayout(input, 0).size()});
-  }
-
-  DALIDataType GetType(const BatchContainer&) override {
-    return DALI_UINT8;
-  }
-
-  void FillOutput(workspace_t<Backend>& ws) override {
-    const auto& input = ws.template Input<Backend>(0);
-    auto& output = ws.template Output<Backend>(0);
-    for (size_t sample_id = 0; sample_id < input.num_samples(); sample_id++) {
-      auto layout = GetLayout(input, sample_id);
-      output[sample_id].Copy(
-          make_cspan(reinterpret_cast<const uint8_t*>(layout.c_str()), layout.size()), nullptr);
-    }
-  }
-
- private:
-  const TensorLayout& GetLayout(const BatchContainer& input, int idx) {
-    return GetMeta(input, idx).GetLayout();
-  }
-};
-
-template <>
-void Layout<GPUBackend>::FillOutput(workspace_t<GPUBackend>& ws) {
-  const auto& input = ws.template Input<GPUBackend>(0);
-  auto& output = ws.template Output<GPUBackend>(0);
-  for (size_t sample_id = 0; sample_id < input.num_samples(); sample_id++) {
-    auto layout = GetLayout(input, sample_id);
-    auto output_ptr = output.raw_mutable_tensor(static_cast<int>(sample_id));
-    cudaMemcpyAsync(output_ptr, layout.c_str(), layout.size(), cudaMemcpyDefault, ws.stream());
-  }
-}
-
-}  // namespace
-}  // namespace detail
-
-template <typename Backend>
-GetProperty<Backend>::GetProperty(const OpSpec& spec) : Operator<Backend>(spec) {
-  auto property_name = spec.template GetArgument<std::string>("key");
-  PropertyFactory(property_name);
-}
-
-template <typename Backend>
-void GetProperty<Backend>::PropertyFactory(const std::string& property_name) {
-  if (property_name == "source_info") {
-    property_ = std::make_unique<detail::SourceInfo<Backend>>();
-  } else if (property_name == "layout") {
-    property_ = std::make_unique<detail::Layout<Backend>>();
-  } else {
-    DALI_FAIL(make_string("Unknown property key: ", property_name));
-  }
-}
 
 }  // namespace dali
